requirements.txt
selenium
beautifulsoup4
google-generativeai
pydantic
pandas
openpyxl
webdriver-manager